requirements.txt
smbus2==0.4.1
configparser==5.2.0
pyserial==3.5
paho-mqtt==1.6.1
pynmea2==1.18.0
requests==2.28.1

adafruit-board-toolkit==1.1.0
adafruit-blinka==8.2.1
adafruit-extended-bus==1.0.2
adafruit-circuitpython-tsl2591==1.3.5
adafruit-circuitpython-mpu6050==1.1.14
adafruit-circuitpython-ahtx0==1.0.14